+++ src/serialization.ml
@@ -285,6 +285,8 @@ let rec value_to_yojson (v : Ast.value) : Yojson.Safe.t =
       ]
   | VFactor _ | VSerializer _ | VBuildLog _ ->
       invalid_arg "value_to_yojson: VFactor/VSerializer/VBuildLog is not supported for JSON serialization"
+  | VNullNode ->
+      invalid_arg "value_to_yojson: VNullNode is not supported for JSON serialization"
   | VUnquote _ | VUnquoteSplice _ | VDynamicArg _ | VQuo _ | VEnv _ ->
       invalid_arg "value_to_yojson: metaprogramming intermediate values are not serializable"
   | VLens l ->